Revamped grading system structure and modularised code
