+++ grading-system/app/services/gcs_service.py
@@ -1,18 +1,14 @@
 import os
 import tempfile
-import fitz
 from google.cloud import storage
-from dotenv import load_dotenv
-
-load_dotenv()
 
 storage_client = storage.Client()
 
-def list_pdfs_in_gcs(bucket_name):
+def list_pdfs(bucket_name):
     bucket = storage_client.bucket(bucket_name)
-    return [blob.name for blob in bucket.list_blobs() if blob.name.endswith(".pdf")]
+    return [b.name for b in bucket.list_blobs() if b.name.endswith('.pdf')]
 
-def download_pdf_from_gcs(bucket_name, filename, local_path=None):
+def download_pdf(bucket_name, filename, local_path=None):
     if local_path is None:
         local_path = os.path.join(tempfile.gettempdir(), filename)
     os.makedirs(os.path.dirname(local_path), exist_ok=True)
@@ -21,26 +17,19 @@ def download_pdf_from_gcs(bucket_name, filename, local_path=None):
     blob.download_to_filename(local_path)
     return local_path
 
-def extract_text_from_pdf(pdf_path):
-    text = ""
-    doc = fitz.open(pdf_path)
-    for page in doc:
-        text += page.get_text("text") + "\n"
-    return text
-
-def upload_file_to_gcs(bucket_name, file):
+def upload_file(bucket_name, file):
     bucket = storage_client.bucket(bucket_name)
     blob = bucket.blob(file.filename)
     blob.upload_from_file(file)
 
-def delete_file_from_gcs(bucket_name, file_name):
+def delete_file(bucket_name, file_name):
     bucket = storage_client.bucket(bucket_name)
     blob = bucket.blob(file_name)
     blob.delete()
 
-def create_gcs_bucket(bucket_name):
+def create_bucket(bucket_name):
     storage_client.create_bucket(bucket_name)
 
-def delete_gcs_bucket(bucket_name):
+def delete_bucket(bucket_name):
     bucket = storage_client.bucket(bucket_name)
     bucket.delete(force=True)
