+++ grading-system/run.py
@@ -1,0 +1,9 @@
+from app import create_app
+from dotenv import load_dotenv
+
+load_dotenv()
+
+app = create_app()
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0", port=app.config['PORT'])
